Improve error checking in Invoke-SIMSreport
I have improved the error checking in Invoke-SIMSreport
It now captures and displays errors from the
commandReporter.exe binary (if any were given)

I have added support for the -PassThru parameter
for Invoke-SIMSreport, the -PassThru parameter causes the
results to be returned directly from the cmdlet.

If -PassThru is provided, a temp file is created and cleaned
up afterwards.
If -Path is provided, that file is used for output.
If both are provided, the file is used, and output is returned.